requirements.txt
async-timeout==4.0.3
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
flask-redis==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-cloud-vision==3.7.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
pdf2image==1.17.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyzbar==0.1.9
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
